Add files via upload
a few of the sample files I've been working on. The material instance creator is from a tutorial series. The asset validator is a project I wanted to work on a while ago and haven't been able to fully figure out but I'm starting to get a much better understanding of how unreal and python within unreal operate.